requirements.txt
#opssdk===0.0.87
SQLAlchemy==1.3.23
shortuuid==1.0.11
boto3==1.26.134
APScheduler==3.6.3
requests==2.28.2
redis==4.5.1
tornado==6.2
PyMySQL==0.9.3
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.25
aliyun_python_sdk_core==2.13.35
aliyun_python_sdk_ecs==4.24.2
aliyun_python_sdk_vpc==3.0.38
aliyun_python_sdk_rds==2.5.11
aliyun-python-sdk-slb==3.3.13
aliyun_python_sdk_r_kvstore==2.1.2
aliyun_python_sdk_polardb==1.8.6
aliyun-python-sdk-alb==1.0.10
aliyun_python_sdk_cdn==3.8.7

tencentcloud-sdk-python-common==3.0.854
tencentcloud-sdk-python-cvm==3.0.854
tencentcloud-sdk-python-cdb==3.0.856
tencentcloud-sdk-python-redis==3.0.885
tencentcloud-sdk-python-vpc==3.0.856
tencentcloud-sdk-python-clb==3.0.873

cryptography==39.0.2
python-redis-lock~=3.7.0
fire~=0.5.0

pyvim==3.0.3
pyvmomi==7.0.3

# 云解析
aliyun-python-sdk-alidns==3.0.3
tencentcloud-sdk-python-dnspod==3.0.867
dnspod-python==1.0.0
GoDaddyPy==2.3.4
#
python-consul2==0.1.5